Blogger/requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==4.2.9
django-crispy-forms==2.1
django-embed-video==1.4.9
django-social-share==2.3.0
idna==3.6
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
